requirements.txt
setuptools
wheel
build
pytest
pytest-cov
setuptools
flake8
black
mypy
Pillow
pillow-heif
pyinstaller
